binds events to dom
